+++ tests/core/config/themedui2.js
@@ -23,6 +23,11 @@ bender.test( {
 		} );
 	},
 
+	// (#4918)
+	test_startup_computed_state_value: function() {
+		assert.isTrue( this.editor.config.useComputedState, 'config.useComputedState should return true as default' );
+	},
+
 	test_startup_focus: function() {
 		assert.isFalse( this.editor.focusManager.hasFocus );
 	},